gke-config-spanner/django-spanner/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
django-storages[google]==1.13
django-environ==0.9.0
gunicorn==20.1.0
google-cloud-secret-manager==2.12.0
drf-yasg
django-google-spanner